requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
beautifulsoup4==4.9.1
brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
contractions==0.0.24
cssselect==1.1.0
dash-core-components==1.9.1
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
dash==1.11.0
distlib==0.3.0
feedfinder2==0.0.4
feedparser==5.2.1
filelock==3.0.12
flask-compress==1.5.0
flask==1.1.2
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0 ; python_version < '3.8'
importlib-resources==1.5.0 ; python_version < '3.7'
inflect==4.1.0
itsdangerous==1.1.0
jieba3k==0.35.1
jinja2==2.11.2
joblib==0.15.1
lxml==4.5.1
markupsafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.4
packaging==20.4
pandas==0.25.3
pillow==7.1.2
pipenv==2018.11.26
plotly==4.7.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pyahocorasick==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pyyaml==5.3.1
regex==2020.5.14
requests-file==1.5.1
requests==2.23.0
retrying==1.3.3
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
textblob==0.15.3
textsearch==0.0.17
threadpoolctl==2.0.0
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.1
tox-pipenv==1.10.0
tox==3.15.1
tqdm==4.46.0
unidecode==1.1.1
urllib3==1.25.9
virtualenv-clone==0.5.4
virtualenv==20.0.21
werkzeug==1.0.1
zipp==3.1.0 ; python_version < '3.8'
